Use appropriate 401 Unauthorized status code when deletion fails due to the wrong base 64 key
The 404 response from a few lines before if there are no results would tell an attacker anyway if the delete request was successful. Another option would be to always return 200 but then we cannot differentiate anything on the client. Let's not obscure too much, the auth code should provide enough security so an attacker can't delete messages